rename folder from 'modules' to 'utils'
